requirements.txt
openai
python-dotenv
fastapi
uvicorn
python-multipart
aiofiles
jinja2
websockets

